requirements.txt
setuptools~=65.5.1
wheel~=0.43.0
Django~=4.2.14
asgiref~=3.8.1
sqlparse~=0.5.0
djangorestframework~=3.15.2
dj_rest_auth~=6.0.0
django-allauth~=0.63.6
requests~=2.32.3
jwt~=1.3.1
django-filter~=24.2